--- src/main/java/com/gae4coon/cloudmaestro/domain/ec2/dto/ApiNameRequest.java
@@ -1,10 +1,0 @@
-package com.gae4coon.cloudmaestro.domain.ec2.dto;
-
-import lombok.Data;
-
-@Data
-public class ApiNameRequest {
-    private String platform;
-    private String instanceType;
-}
-
