--- src/main/java/com/gae4coon/cloudmaestro/domain/ec2/dto/requestPrice.java
@@ -1,13 +1,0 @@
-package com.gae4coon.cloudmaestro.domain.ec2.dto;
-
-import lombok.Data;
-
-
-
-
-@Data
-public class requestPrice {
-    private String Platform;
-    private String Instance;
-    private String LifeCycle;
-}
