+++ src/main/java/com/gae4coon/cloudmaestro/domain/price/dto/ApiNameRequest.java
@@ -1,0 +1,10 @@
+package com.gae4coon.cloudmaestro.domain.price.dto;
+
+import lombok.Data;
+
+@Data
+public class ApiNameRequest {
+    private String platform;
+    private String instanceType;
+}
+
