+++ src/main/java/com/gae4coon/cloudmaestro/domain/price/dto/rdsCostRequest.java
@@ -1,0 +1,9 @@
+package com.gae4coon.cloudmaestro.domain.price.dto;
+
+import lombok.Data;
+
+@Data
+public class rdsCostRequest {
+    private String dbEngine;
+    private String instanceType;
+}
