+++ src/main/java/com/gae4coon/cloudmaestro/domain/price/dto/requestPrice.java
@@ -1,0 +1,13 @@
+package com.gae4coon.cloudmaestro.domain.price.dto;
+
+import lombok.Data;
+
+
+
+
+@Data
+public class requestPrice {
+    private String Platform;
+    private String Instance;
+    private String LifeCycle;
+}
