--- src/main/java/com/gae4coon/cloudmaestro/domain/rds/dto/rdsCostRequest.java
@@ -1,9 +1,0 @@
-package com.gae4coon.cloudmaestro.domain.rds.dto;
-
-import lombok.Data;
-
-@Data
-public class rdsCostRequest {
-    private String dbEngine;
-    private String instanceType;
-}
