+++ src/main/java/com/gae4coon/cloudmaestro/domain/rehost/dto/Ec2.java
@@ -1,0 +1,4 @@
+package com.gae4coon.cloudmaestro.domain.rehost.dto;
+
+public class Ec2 {
+}
