--- src/main/java/com/gae4coon/cloudmaestro/domain/standard/dto/Ec2.java
@@ -1,4 +1,0 @@
-package com.gae4coon.cloudmaestro.domain.standard.dto;
-
-public class Ec2 {
-}
